apollo-ai-m2/pyproject.toml
[project]
name = "apollo-ai-m2"
version = "0.1.0"
description = ""
authors = [
    {name = "solideos",email = "urim.chae@solideos.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "easydict (>=1.13,<2.0)",
    "pandas (==2.1.4)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "pymilvus (>=2.6.6,<3.0.0)",
    "scipy (==1.11.4)",
    "pyarrow (==14.0.1)",
    "nvidia-cublas-cu12 (==12.1.3.1)",
    "nvidia-cuda-runtime-cu12 (==12.1.105)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
